Enable a mode where the same process is running multiple engines
This greatly reduces the number of processes that we have to spawn,
which reduces our memory footprint and lets the GC do a better job.
For users with a developer key, we still run them in a private process
because they can upload arbitrary code.

Or actually, all users are still run in separate processes until
we test this better.