Implement new way to upload and dowload devices
Instead of having just one checkbox that says "i have a js package or
not", have a dropdown with multiple choices. One of the choices
is the custom JS package, but other choices allow declarative definitions
of the devices with different formats, such as RSS, SOAP or different JSON+REST
conventions.