End the conversation if a command is an action.
By default, google home will keep listening if we don't end the conversation unless the user says "stop".
We keep the conversation if the command is a "get", so one don't need to say "talk to almond music" every time while asking multiple questions.
We end the conversation when the command is a "do", so music will starts playing right away.