Add automatic link checking
Crawl the website and poke all the links that we find recursively
from /, looking at broken links or pages where the server dies.

We do this both from outside users, from regular logged in users,
and from root, so we test all different paths that we can take.

This does not replace functional testing, because it does not
check anything beyond HTTP 200, nor selenium integration
testing, because it doesn't actually anything shows up in the
browser, and it does not check images or CSS.