Clarify terms of service wrt copyright of submitted contributions
Basically, if there is no written agreement between you and us,
and you make a contribution to Thingpedia, you give us a full license
to do anything to Thingpedia, or at least a GPL-compatible license.
If you cannot do that (eg you are submitting code that is under a
GPL-incompatible license), then you can't submit at all, and the
submission will be violated.

If someone's right are infringed by a submission, they should send
us a DMCA takedown notice (that way, we don't have to police submissions
proactively, and we can just wait for people to complain)

If there is a written agreemeent (eg a research sponsorship agreement,
a separate licensing deal, or the MTurk worker agreement), then that one
wins.