Fix tests
Hidden among this changes in this branch is a small change to the
canonical form of the @messaging kind. This causes the
synthetic command to change from "get commands of messaging" to
"get commands of messaging interface".
In turn this means we need to update the expected hashes of the
dataset.