example_utterances: avoid full text index for search
Instead, just use a regular expression match.

In theory, this is slower, as it has to scan all rows that match
the rest of the query. In practice, it is much fast to scan
through all rows in the cheatsheet (type = 'thingpedia', a couple
hundreds at most) than to scan through all rows that match the
fulltext index (which includes synthetic and replaced sentences,
so it is potentially hundreds of thousands of sentences).

In numbers:
- with regexp: about 0.01 s ~ 0.04 s
- with fulltext: 9.5 ~ 11 s