EngineManager: handle child being disconnected
Otherwise a race can occur if the web process attempts a direct
connection between the child disconnecting and the process dying.
If the direct socket handle is sent quickly enough, the send
will go through even though the process is about to die.

This will corrupt nodejs' accounting of sockets sent to child
processes, and cause net.Socket.close to never terminate.
This in turn results in an unkillable master process if any
child is dead when the master process is taken down.