conversation: wait until the first message to declare the socket is ok
If the engine is dead, we will successfully open the websocket,
but the websocket will be closed immediately afterwards.
This throws off the backoff calculation and causes us to keep trying in
a very tight loop.
Instead, wait until we get "Welcome back!" to declare the engine is
alive. If it is not, we will increase the backoff timer exponentially
(until the user refreshes the page or brings the engine back to life).