Remove the distinction between "adding" and "suggesting" a command
At the front page level, before the user is familiar with Thingpedia
and ThingTalk, it's hard to know what's adding and what's suggesting.

Instead, let all users go through Add. We semantic parse the command,
and show the choices to the user. If none are good, the user will
hopefully click "None of the above", at which point we give them
the option of suggesting the command without writing ThingTalk for it.