home: change "Create New Device" to "Explore Thingpedia"
Prompting the users to create a new device before they have
looked at what devices are available is wrong, and the Commandpedia
is only a small subset of all supported devices.

Instead, we should have them look at Thingpedia (or use search)
first, and then, if that fails, go through the process of becoming
a developer and contributing a new device.