home page: remove target_code hover from commandpedia commands
target_code uses NN syntax rather than surface syntax, and the
fact that it is slightly different (eg uses `param:`) plus has
constants replaced with numbered tokens can be confusing to
developers.

The correct way would be to reconstruct surface syntax from NN
syntax, but that would require having the original constants
from the sentence, which we don't store because we don't need.