Add #[package_version] annotation to classes when imported
In the future, this will allow not to store a new package when
the user is not uploading new code.
Unfortunately, the future is not now: package_version and version
must be identical to be compatible with clients that don't respect
package_version. The future will happen when those clients go away.