Fix optimization to avoid updating the dataset if not needed
Checking the examples one by one is not enough because it won't
account for collapsing.

Instead, do a quick check by comparing the source code first.

This optimization is important as it avoids locking rows in
the example_utterances table, which will cause the transaction
to queue behind the update dataset task.