Implement public profiles for organizations and users
In the organization's profile, we show a list of members and the
list of devices contributed by that organization (that have
been approved).

In the user's profile, we show a profile picture of the user (using
Gravatar and a hash of the user's email) and the user's real name.

The user must opt-in to be listed publicly, in their organization,
and must opt-in to have their name and profile picture visible.