training: move replaced dataset to a separate table
Having replaced (blown-up) sentences in the example_utterances
table is creating a lot of problems with locking and contention
with other accesses to the table (eg concurrent editing of thingpedia
devices).

There is no need for the two datasets to be in the same table,
as in practice one always wants to operate on the quoted version
of the data (in example_utterances) and use the replaced data
(blown-up, quote-free) only for training.
Contention for access to the new table is also not a problem
because the training service runs only one dataset update at a time.