user utils: merge redirectLogIn and requireLogIn
The distintion is whether the request was a GET (clicking a link,
to which it's fine to respond with a redirect) or a POST/PUT/DELETE/etc.
(where a redirect is dangerous and can cause the data to be retrasmitted,
so we should simply return an error).
We can check the request method to learn that, and simplify the API.