Don't logout the user when they navigate to /user/login
Because user information is initialized earlier, this can lead to
a confused state where the UI appears to be logged in (eg.
Settings, Log in, Admin items are shown) but clicking on them
redirects to /user/login again.

Instead, just redirect to / if the user tries to log in again.