training/update-dataset: delete existing dataset in a separate transaction
This ensures that we release the locks acquired by the delete
operation early enough, and we don't block concurrent reads to
the example_utterances table.
We cannot unfortunately rely on the granularity of database locks
for concurrency (even though read/writes don't ever overlap,
cause reads are only for real data and we operate only on generated
data in this script) because in different code paths we access the
table through different locks (schema_id, language_type, language_flags).

Luckily, provided that two instances of update-dataset are never
run in parallel, it is safe to drop the locks early: update-dataset
is the only place where we change the set rows that can materially
be affected by the delete operation (regardless of what the db's locking
thinks), hence there is no race with concurrent writes by the
website.

A race *can* occur between update-dataset and download-dataset
(as run by train). The probability of it occurring is minimal though:
the train job has to start exactly in the middle of the update-dataset
job; normally, a train job follows an update-dataset job, so it is not
an issue. Also, we can always reissue a train-only job when there is
no update-dataset and bring back consistency.