Add a script to do ppdb augmentation/parameter replacement by hand
For research purposes, the automatic training scripts are "too efficient":
they are black boxes that are tied to the production model, while
often we need to open the box and tweak the parameters. This is
far easier with files than with SQL, so this new scripts operates
on standard-format TSV files (id, sentence, program).