Implement persistent Commandpedia likes
Add a new column `like_count` to example_utterances, and a new
table tracking likes for each user.
Duplicate likes are now ignored, and the user must be logged in
to like an example. If you click on a like and you are not logged
in, you are redirected to a login page.