Add email verification functionality
Add new field to the users table: email_verified

The field is false for new users, unless the users come from
"Log in with Google".
Users are restricted from certain functionality while their email
is unverified.

A verification email with a link/token is automatically sent
upon registration, and can be resent from the profile page.