Implement 2-factor authentication
Based on the TOTP standard

Compatible with Google Authenticator and Duo Mobile.

2FA is optional and can be enabled from the user settings page.
If 2FA is enabled, the user must enter a TOTP code after login
(on a separate page) before he is considered authenticated.
If 2FA is not enabled, login works as before.

If 2FA is enabled, the user must also enter a TOTP code to reset
his own password.
Resetting 2FA is not implemented (must be done manually).