Implement authentication of the master socket
This way, developer sandboxes cannot try to connect to the master
socket, even if that is exposed over TCP.

The authentication token is sent unencrypted, as it is assumed
everything is deployed in a trusted network.
A bigger refactoring that eliminates direct sockets (and cut down
transparent-rpc usage, which is leaky and problematic) would be needed
to enable encrypted frontend<->backend communication.