Convert ADMIN developer status to THINGPEDIA_ADMIN role
The current developer status field is ill-defined:
- tester & developer levels are specific to an organization
  (read-only vs read-write access)
- trusted & admin levels are global

Practically, an organization admin should be able to toggle
between tester & developer on any org member, but only a thingpedia
admin or site admin should have the power to grant the trusted
bit or the admin bit.
This is the first step in this direction.