Fix error message when doing org-admin operations as regular user
If a developer user (not an org admin) tries to do an org-admin
operation, we should show an error message, rather than sending
him to the developer application form.

In fact, we should show an error message even if a non-developer
user tries to do a developer or tester thing. We already hide
links appropriately (linkchecker would barf on the 403 otherwise),
so we don't need this special handling.