sandbox: allow disabling journal redirection
In Travis, we want to see the error messages from the child
processes on stdout/stderr, otherwise we never notice bugs
and we cannot debug.

In fact, right now Travis is failing with
User 1 failed to start: Exited with code 1

but nothing else fails the tests and we wouldn't notice
unless we paid extra attention.