tests: flip around which test is run with & without sandbox
"bob", the user we use for the web almond tests, is not a developer
inside webalmond-integration.sh, but is a developer for
thingpedia-integration.sh
That makes is unsuitable to test the sandbox - as we'd test
a non sandboxed user instead.

To make sure we actually test the sandbox, make sure that bob
is sandboxed.