sandbox: support multiple prefixes for almond
The node binary might also be installed in a weird location
(eg under a home directory) outside the /usr hierarchy.
(This is the case in Travis, which uses nvm)

In that case, bind-mount the node binary inside the sandbox.
This setup assumes that node is linked statically against
its own libraries (as it is when installed from nvm or
the nodesource repositories), so the only shared libraries
are the C/C++ runtimes