Add input validation module
Add a middleware to all routes that validates that query/body
parameters have the expected type.

This reduces the amount of accidental garbage that can creep
into our SQL queries, and in turn can avoid 500 errors due
to bad SQL syntax, or other bugs due to inconsistent stringification
of objects/arrays that maliciously appear in the query.