enginemanager/controlsocket: don't wait for the server to close
If a socket is sent to a subprocess, and that subprocess is killed
hard (eg by /admin/users/kill/all or a "kill" shell command), nodejs
will become confused and lose track of the socket. In that case,
socket.close() will hang, and prevent us from closing the process
gracefully.

We already know that the process that received the socket died, hence
the socket cannot be leaked and the connection was shutdown (hopefully
gracefully), so we can just keep going and close the master process
too.