routes/my_api: fix botched handling of CORS
Previous code allowed any origin to issue requests, provided that
they included an Authorization header - even if that Authorization
header was garbage, because it was not checked later.

Now instead we let the request go through with cached (session/cookie)
credentials if the Origin is truly ok, and otherwise require
valid OAuth authentication.