Remove incorrect test
We have to rely on the fact that the browser will not issue
a cross-origin request without Origin header, or if they do
it will not have credentials.

(Indeed, it seems they do. The wording of the spec is confusing,
see:
https://fetch.spec.whatwg.org/#concept-http-network-or-cache-fetch
point 5.10
and:
https://bugzilla.mozilla.org/show_bug.cgi?id=1508661

I don't think that's correct based on the spec writing but
apparently there are tests that expect that behavior, which
means there is weird web-compatibility going on...
)