Make use of genie-toolkit to augment the dataset
This requires a small refactoring of how the augmenter is invoked,
as Genie provides an interface based on Stream.Transform.

To complement that, we add a database writer in the form of
Stream.Writable (with builtin batching/multi-insert), and
two helper Transform streams to manipulate type/flags for
synthetic examples and to filter the stream for incremental
updates.

Handling of paraphrases does not use the stream interface as
mysql does not like leaving the cursors open for too long,
and node-mysql will end up buffering the whole resultset
anyway the moment we start doing concurrent queries (such as
loading a parameter dataset).