Add NLP inference server
This used to be written in Python and provided by genieparser,
because it would load tensorflow in-process.

Nowadays instead we can talk to decanlp over RPC, and the server-ish
parts of the server can live in here in JS.

These include:
- interacting with the db to log sentences, do online learning,
  and managing commandpedia
- talking to the tokenizer (which is something many parts of genie
  need too)
- reloading when the training service says so

It's better for all of these to live here, so that configuration
is in one place, and DB migrations and internal API changes can be
managed atomically.