Add Trie data structure
A Trie is a data structure for efficient storage of the key value
pairs where the key is a sequence (e.g. a string or sentence).

This is a naive implementation of a Trie without path compression,
iteration, or deletion, which is ok for our needs and keeps
the implementaton simple.