Strip regexp-unsafe tokens from example search keys
Example uses a manually constructed regular expression for search, because
the fulltext index is too slow to keep updated.
We need to be very careful in how we construct this regular expression
though, as the regular expression is under user control and we must
remove any unsafe character.

We remove rather than escaping the characters because the processing
happens after tokenization, and dangerous characters have been split
into their own tokens already.