nlp/main: use non-locking reads when loading the exact matches initially
The whole transaction is readonly, and we use a transaction only
to see a consistent snapshot. "repeatable read" isolation level
will use validation (versioning) instead of read locks, which
ensures we don't block and timeout when the training server is also
modifying the exact matches.
Note that for readonly transactions, "repeatable read" and "serializable"
are semantically the same (just not operationally).

(On reload, we don't use a transaction at all, which keeps the default
of "repeatable read")