Import device: approve the existing schema when uploading an auto-approved device
If the user updates a device and clicks on "Approve automatically"
but also does not modify the schema (e.g. only the code or the dataset
are modified) we would leave the schema unchanged and potentially
not approved.
This could lead to a database inconsistency, where the device is
approved but the schema is not.

Fixes #236