util/db: wrap callback result into a Promise before .then
If the callback passed to withClient does not return a promise,
we would crash with a type error accessing .then(). Instead,
treat it like a promise that resolved immediately.

This happens in particular with reloading exact matches for non-default
models, as we open a connection but then call the dummy exact matcher
that does nothing with it, and returns undefined immediately.