Stop using generic Error class
Introduce an error hierarchy for all errors that we throw explicitly.

These errors are different than library or programming errors,
because 1) the user can cause them at runtime, 2) we should show
them to the user, 3) they are associated with a specific HTTP status.

For completeness, and for ease of `grep 'throw new Error'`, we also
introduce the InternalError class, which provides useful error codes
for debugging.