training: separate generating a training set from updating the exact match dataset
The exact match dataset has to go in MySQL because that's the easiest
to pull from the NLP server, and it also makes sense to update
incrementally. OTOH, the training set is much larger, which puts strain
on MySQL, and updating it incrementally is not as useful, especially
when the bottleneck is generating the synthetic part (which is not
incremental). Instead, regenerate the training set from scratch
every time a model is trained - training does not happen often,
and the cost is absorbed by the training procedure anyway.