alexa: use full thingtalk syntax instead of NN syntax + slot objects
SLOT_ don't survive JSON serialization. Turns out that for
this reason almond-dialog-agent has a HACK that converts from JSON
serialized Ast values, but that loses a lot of information (for
example the unit of duration). Instead, just serialize to full
ThingTalk syntax, which loses no information and is easier to handle.