Sanitize zip file names
This is another false positive, but it's a very hard case to prove:
all names are valid ThingTalk class identifiers, which consists
exclusively of non-empty alphanumeric sequences, sepated by '.',
and therefore are valid filenames.

At the same time, there are enough paths calling these functions
that I am not completely comfortable with just a false-positive mark.
Instead, let's be defensive and sanitize the input.