i18n: normalize traditional vs simplified chinese correctly
We need to preserve the distinction of the two Chinese variants
in our datasets, models and template packs.

In the future, we need to be a lot more thorough in handling
country locale variants even for English (e.g. metrics vs Imperial,
country currency, etc.) but for now, this suffices.