training: don't swallow errors from generation
If generation fails inside the sandboxed worker, the child process
will exit with status 1 (unless there is a crash, e.g. OOM). We
need to detect that and fail the piped streams, otherwise node
will just happily close the read stream and we'll pretend we finished
successfully, potentially with a truncated dataset.

Fixes #397