util/upload_dataset: handle CSV files with three columns correctly
To roundtrip a dataset downloaded from Thingpedia and upload it again,
we need to handle the case of having 3 columns, and distinguish value
from preprocessed.

The cases are:
- 1 column: either value or preprocessed, depending on the `tokenized` flag
- 2 columns, the second is a number: either value or preprocessed (depending
  on `tokenized` flag), and weight
- 2 columns, the second is not a number: value and preprocessed, weight is 1.0
- 3 columns: value, preprocessed, and weight

If 3 columns are provided, the `tokenized` flag is ignored.

This fixes accidentally putting uppercase words in the programs, and
messing our location models dramatically.