training: store job information in MySQL
Instead of keeping dependency and queues in memory, and using
a local file for persistency, store all the training jobs (previous,
current and queued) in mysql. This avoids depending on the local
filesystem (which might not be persistent at all) and in the future
should allow us to make the training daemon truly stateless,
including the ability to die and restart without killing existing
jobs.