mturk: add validation support
Add status tracking to mturk batches, to distinguish the paraphrasing
from the validation phase. When the validation phase starts,
we create validation HITs using code borrowed from Genie (which
should be shared). The validation HITs can be accessed on the website
in a manner similar to the paraphrase HITs, and the results are
stored in the database immediately.

Validation operates on the `training` flag of example sentence
in the database, which controls whether the sentence is included
in the training set or not.