training: fix preparing the training set
The whole prepare-training-set task is run inside a single database
transaction, to ensure a consistent snapshot of Thingpedia is used.
At the same time, most of the queries are run at the beginning
of the transaction only. This can cause the server to close
the connection due to inactivity, which later would cause us
to fail when committing the transaction.

Fixes #517