mturk: wrap errors during entity allocation correctly
If the entity is a generic entity, or a location, we need to pass
the sentence to NN to identify it. Also, if an entity is the present in
the program but not in the sentence, toNN will fail. We need to
catch that and show an error to the user.