registration: do not talk about a verification email if one was not set
In Login with Github, we might sometime not receive an email from
GitHub. In that case, we don't send out the verification, until
the user sets one from the Settings page.

Yet, if we say that there is an email when there isn't, the user
might be tempted to click on Resend Verification, causing us
to send an email to a null address and crash.