training: pass tensorboard path directly to decanlp if possible
decanlp has added an option to save tensorboards in a different
folder, and we are switching to EFS instead of S3 to store tensorboard,
which is faster and cheaper. With this patch, decanlp will write
directly to the shared tensorboard directory on EFS, which means
writes update in real time and we only transmit the delta and not
the whole file.