boostrap: create the SQL table during bootstrap
Instead of having two separate steps, execute-sql-script and then
boostrap, execute both as part of the same bootstrap step. This
is easier for users setting up Almond Cloud for the first time.

To avoid mistakes wiping the whole database, we now check if the
root user has been created (indicating bootstrap has run before)
and do nothing in that case. A new argument, --force, can be passed
to force-wipe the database (this is used in tests).