nlp: temporarily remove the frontend classifier
The classifier is brittle, crashes often, is not very good at its
job, and it's unused anyway (because we don't know what to do with
questions or chatty input).
Let's remove it, and if we want it, we can bring it back later.

Note: the comment in the dockerfile says PPDB too, but PPDB has
not been in the image since ed256f6326fb8f147530d818d9c95ed49345c603.

Fixes #656