cloud-sync: fix race condition resulting in stuck connections
We must begin processing messages on the websocket as soon as we
receive the websocket connection, rather than after retrieving
the user from the database, otherwise some messages might be
dropped. To do so, we need to buffer all incoming messages until
the backend is ready to process them.
(In the future, it will make more sense for the server to send
a "hello" message in the web socket first. But this is what we
have so far. Lessons learned!)