Add the ability to generate and download custom datasets
Add a new set of training job types, which run the dataset
generation and save the output in S3, and add a new UI in the
developer console to run these jobs. Any developer with sufficient
credits can start a job, wait a while, then come back and
download the dataset.