tests/nlp: make the tests actually useful
Test that, for all the sentences that we try, the server responds
a valid program.

The program is not always correct, but it's correct most of the time.
When it's not correct, it's due to bad tokenization, because we
use the mock-tokenizer which is hacky and buggy. This commit
makes it less buggy, not quite.