Fix querying setup information for Home Assistant switches (#942)
The class org.thingpedia.iot.switch can be configured in two ways:
- by configuring io.home-assistant, which can directly provide
  an org.thingpedia.iot.switch
- by configuring io.home-assistant, which can provide an
  org.thingpedia.iot.light-bulb, which is a kind of
  org.thingpedia.iot.switch

Indeed, this was visible because io.home-assistant was returned
in both branches of the SQL query. Using `union distinct` solves
the issue.

Additionally, we need to return the name of the device being configured
when we return multiple choices, or the client will be confused.