Include examples from canonical annotations in the cheatsheet
Most classes these days do not have any dataset.tt, and rely on
canonical annotations exclusively. To ensure we have something
to show for these in the cheatsheet, make up examples in
the example_utterances table. These examples have the synthetic
flag to exclude them from most APIs, but are included when
importing the device (to update the table), in commandpedia and
in the cheatsheet.