Always set a type for examples in the cheatsheet
It's a bit weird when some have examples have WHEN/GET/DO and some
don't, and it's hard to understand it has to do with the part of
speech. Instead, identify the function type (with a bit of hack
to avoid loading the class def) and set a type that way.