Remove Train Almond
It does not work with dialogues, and the data goes nowhere because
it goes into the MySQL db instead of train/dev sets we normally use.

People should use commandpedia to submit commands that they would
like supported, and should talk to us on Github/community forum
to add new features to skills.