Add a SQL-backed implementation of Tp preferences class
The existing file preferences relies on persistent local disk storage,
which we want to get rid of because it conflicts with the design
of the Kubernetes backend.

Worker initialization is a bit more complex now because platform
initialization is asynchornous, and we cannot construct the Engine
until the platform is fully initialized.