dbproxy: implement authentication
Have the master process issue a JWT token which is passed to the
worker in the engine configuration. The token is included by the
worker in all requests to dbproxy, which validates and extracts
the user ID. The user ID is no longer included in the URL,
because it's now part of the access token.