K8s-backend: update user state when backend is removed
Fixes #1086
To update a User CR, we need to get the User object from k8s api call.
Previously handleSharedUser will return a nil user if backend or engine
status is not available.  This will cause the User state not updated
when the backend pod is removed.  The fix is to always return the User
object if it's available from k8s api so all the errors are propogated
to the User state.